+++ Lykke.CoinMarketCap.Client/Models/CryptoCurrency/Quote.cs
@@ -6,19 +6,19 @@ namespace Lykke.CoinMarketCap.Client.Models.CryptoCurrency
     public class Quote
     {
         [JsonProperty("price")]
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
         [JsonProperty("volume_24h")]
-        public double Volume24H { get; set; }
+        public decimal Volume24H { get; set; }
 
         [JsonProperty("percent_change_1h")]
-        public double? PercentChange1H { get; set; }
+        public decimal? PercentChange1H { get; set; }
 
         [JsonProperty("percent_change_24h")]
-        public double? PercentChange24H { get; set; }
+        public decimal? PercentChange24H { get; set; }
 
         [JsonProperty("percent_change_7d")]
-        public double? PercentChange7D { get; set; }
+        public decimal? PercentChange7D { get; set; }
 
         [JsonProperty("market_cap")]
         public long MarketCap { get; set; }
